feat(scrolling-list): split scrolling list controls buttons
