+++ packages/components/carousel/src/Carousel.doc.mdx
@@ -144,6 +144,16 @@ The `unstyled` prop of `Carousel.PageIndicator` allows you to use your own style
       description:
         'Two interactive elements, styled as arrow buttons, that displays the next or previous slide in the rotation sequence.',
     },
+    'Carousel.PrevButton': {
+      of: Carousel.PrevButton,
+      description:
+        'Button that scroll to previous set of slides.',
+    },
+    'Carousel.NextButton': {
+      of: Carousel.NextButton,
+      description:
+        'Button that scroll to next set of slides.',
+    },
     'Carousel.PagePicker': {
       of: Carousel.PagePicker,
       description:
